+++ client/app.js
@@ -4,6 +4,9 @@ let sketch2Folder = 'perlin/';
 let sketch3Folder = 'code-rain/';
 let sketchName = 'index.html';
 
+let currentSketchIndex = 1;
+let sketchContainerName = 'sketch-container';
+
 let sketches = [
 {
 	name: 'blobs',
@@ -21,10 +24,45 @@ let sketches = [
 }
 ];
 
-sketches.forEach (function(sketch) {
+displayOnly(sketches[1]);
+
+// sketches.forEach (appendToDOM);
+
+function appendToDOM(sketch) {
 	let iframe = document.createElement('iframe');
 	iframe.src = sketch.path;
 	iframe.width = window.innerWidth;
 	iframe.height = window.innerHeight;
-	document.getElementById('sketch-container').appendChild(iframe);
-});
+	document.getElementById(sketchContainerName).appendChild(iframe);
+}
+
+function removeAllSketches() {
+	let sketchContainerNode = document.getElementById(sketchContainerName);
+	while(sketchContainerNode.hasChildNodes()) {
+		sketchContainerNode.removeChild(sketchContainerNode.lastChild);
+	}
+}
+
+function displayOnly(sketch) {
+	removeAllSketches();
+	appendToDOM(sketch);
+}
+
+function displayNextSketch() {
+	if (++currentSketchIndex == sketches.length) {
+		currentSketchIndex = 0;
+	}
+	displayOnly(sketches[currentSketchIndex]);
+}
+
+function displayPreviousSketch() {
+	if (--currentSketchIndex < 0) {
+		currentSketchIndex = sketches.length -1;
+	}
+	displayOnly(sketches[currentSketchIndex]);
+}
+
+let prevButton = document.getElementById("previous");
+prevButton.onclick = displayPreviousSketch;
+let nextButton = document.getElementById("next");
+nextButton.onclick = displayNextSketch;