+++ CHANGELOG.md
@@ -4,6 +4,8 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/).
 
 ## [Unreleased]
+## Added
+- Next/previous sketch buttons.
 
 ## [0.4.0] - 2017-05-10
 ## Added